services/common/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
python-dotenv==1.0.1
cryptography==42.0.8
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
httpx==0.27.0
